Create a fake user in the dummy app
This dummy app's User model is an in-theory stand-in for what will, in practice, be the engine users' own user models. Those user models will more often than not be called "User", but why assume?

The mattr_accessor in the Thredded module allows us to configure this user class from the context of the parent application.

For example, in your parent application's `config/initializers/thredded.rb`:

```
# Thredded.user_class = 'User'
Thredded.user_class = 'MyUser'
```