Topic spec passing and cleaned up a little bit
The topic spec had a few things going on in there that didn't need to be
in there. There were several Ability related tests that should be in the
ability spec, so those have been extracted out to there for more focused
testing.

In order for topic tests to pass there is the dependency that a user
object exists and is configured in the dummy app. Creating the thredded
initializer in the dummy app and pointing it to the "fake" User got
things passing.

Post model and spec are not passing but several things in there were
cleaned up and all were wrapped in the correct Thredded module.