[JRO] Use config.to_prepare instead of after_init
config.after_initialize will run once in development mode, not every
time. In production this isn't a problem because the class is cached but
in dev mode this is an issue if whatever you're doing in the block needs
to persist across requests.

In this case it was applying/including a decorator module to a class
just once.

config.to_prepare will run in dev every request.