[JRO] Communicate eager loading of user_topic_reads
It is not obvious that the topic we're pulling for the posts index is an
object that has the user_topic_reads eagerly loaded. If we don't eager
load them then the query will be deferred until AFTER the
user_topic_read for that topic is created, which isn't an accurate
representation of the topic at the time we want it. We want all of that
information about the topic, and it's relationship with
user_topic_reads, before any writes are made regarding that
relationship.

This commit ensures that it's obvious that we're making that query with
the eager load.

Also:

* rescue the ActiveRecord::RecordNotFound error and raise our own TopicNotFound exception.
* Remove the filter select from the post form partial. That's inconsequential now that we've extracted that