[JRO] main app expected to provide thredded-specific layout
The main parent application should define what layout the thredded engine should be rendered within. Up until now it was passing up the chain to whatever implied layout is set in the main application. That approach felt a little dirty so this give the most flexibility and transparency as a contract between the engine, and the parent application mounting it.