[JRO] Rm attachments and translate `t:img` tags to image tags
The current way where we add attachments to a post via the nested attributes -- posts having many attachments, etc -- is a poor method of adding images to posts. As a result the app should allow external methods of doing this. In my mind this should be another engine handling file uploads and displaying it.

So - this commit is the first step to removing these. Instead of completely removing attachments for now we'll leave them in there and will, instead, translate all `[t:img]` tags to real image tags (in markdown or bbcode)

* Remove `accepts_nested_attributes_for :attachments` from Post.
* Add a rake task -- `thredded:update_legacy_timg_tags` -- to make the translation in each post.
* Add two traits for post - `:markdown` and `:bbcode`
* Remove the attached image filter
* Remove attached files form fields
* Remove ability to attach images via email