[JRO] Fix counting of unread private topics
The key used for caching was taking into account only the user, and not
the messageboard it was in as well. A user's unread count in one
messageboard can, and will, be different from the count in the other
messageboard - therefore the cache key for each board and user needs to
be unique.