[JRO] Fix bug w/topic finder
The find_by_slug method name is a little too rails'y and easy to mistake
as being a method owned by the rails api. Renaming it to something a
little more descriptive will mitigate the issue I just had where I
replaced it thinking it was a rails method - but it wasn't.

* Add a `!` to the end of the method to intimate that it might raise an
  exception if a record isn't found
* Remove the `!` from `update_read_status` ... no reason for it, really.