Provide default JS and Sass
Set up the engine to provide default JS and CSS.
For now, only added support for select2 and time ago.
On the CSS side, there is only a default bootstrap-sass theme for now.

This is all integrated with `rake dev` dummy app.

Perhaps Bootstrap styles can live in a separate gems in the future, but
I've just put it here for now.