[JRO] Allow existing user to be a superadmin #66
Allowing any old person to create the first forum is probably a "bad
idea". This commit now introduces the capabilities to assign a user, via
a rake task, the role of a "superadmin". These super-people will then,
eventually, be the only ones able to create a new forum, either at
first, or in the future - period.

Up until now forums can only be created once you JUST install the engine
(or unless you dig into the console and do it manually) and that's just
dumb.

Edit README w/updated ordered list & authorization. The ordered list in
the README was not incrementing correctly so some changes to the text
(indenting properly) fixes that. Also, moving the instructions for
authorizing a user as a superadmin were moved below to the "getting
ready" section of the readme. It's not really an installation dependency.