[JRO] Wrap AR transactions in `with_connection`
When the AR models are updated in a thread context, instead of a
background job manager like resque or sidekiq, the thread doesn't
release the connection to the database - therefore shit gets fucked.
The changes here will wrap them in `AR::Base.connection_pool.with_connection`
which manually obtains, and releases, the db connection after the block
is performed.

Other changes are related to fixing some broken tests.