[JRO] Add sass gem dependencies to parent app
... instead of automatically having them required IN the gem.

This way there is the option to not lock into one sass framework like
bourbon/neat and allow themes to be created using something like
bootstrap or zurb foundation.

Currently the option to pick a sass framework is hidden, and defaults to
bourbon, but the underlying functionality exists for when those other
themes are created.