[JRO] Not finding messageboard is not Exception'al
Current situation:

1. you call `messageboard` from a view
2. `def messageboard` looks for it via the slug in the params hash
3. it has no param, or that messageboard passed in via param does not exist
4. it says "whoops. doesn't exist. I'm raising an exception"

This causes an issue when you're at the root of the thredded engine and your layout uses some code that's looking for a messageboard ... resulting in an infinite redirect loop.

Bummer.

This PR changes that behavior to only return nil on that scope and let people deal with it accordingly where it is being called.