[JRO] User posting automatically makes them member
When a user creates a new topic (private or public), and they were NOT a
member of that messageboard, we should add them to the messageboard as a
member immediately.

Why? In the case of someone posting a private topic, they would not be
able to see it because you must be a member to see it. So - let's give
them the appropriate role and let them see things.

When creating a topic, well -- obviously they want to be a member of the
board at this point. This is an easy and fast way to do this.