[JRO] Toggle active users div on over/out/touch
https://github.com/jayroh/thredded/issues/86

Thought about just using css to toggle this div on mouseover/out but
reconsidered when thinking about the mobile experience. Using javascript
we can more accurately manipulate the behavior based on the various
events we have available.

The css and how it looks needs some work still but this will do in a
pinch and as a starting point.