[JRO] Make sure all users who post are made members
If a messageboard allows a user to post in there, and the user is not a
member yet, when they post then they are made a member of the board.
This allows users to be responded to (via @ replies) and sent private
messages as soon as they are a part of (members of) a messageboard.