[JRO] Copy views/assets to parent /vendor
Mixing the THREDDED files with the parent app files poses some particular issues and is considered, in general, to be a no-no. As such, moving things to /vendor is both accepted and much less of a headache.

For example - copying js and css assets to the parent app can be a huge pain in the ass if the parent app is using the `require_tree` directive in their asset manifests. That would include all of the thredded assets with their own assets and ... ugh, what a damn mess.

Also, if the parent app IS copying over the assets, then we'll need to inject the bourbon and es6 dependencies to their Gemfile so everything works. Natch.

* Adjust the newly namespaced image paths in `assets.precompile` to include the `./thredded/` dir.
* Update README w/new details