[JRO] Small fixes/optimizations for 2 shell scripts
The create-db-users script does not take into account the fact that the postgresql daemon may be running as another user account on the system so this commit defaults it to `postgres` but allows being overridden by the PG_DAEMON_USER environment variable. This use case is semi-common if a developer is using something like [Postgres.app][] from Heroku.

Both `script/migrate` and `script/create-db-users` were logging out to STDOUT with escape codes that were not being interpreted, so this commit also adds descriptive ENV variables denoting that they will come out colored green.

`database.yml` was defaulting to the postgres port and would not work if using mysql, so pulling those into variables in the yaml file makes it a little easier to discern what the values are that are being set.

Lastly, some accompanying README edits to make things more clear (possible options, etc).

[Postgres.app]: http://postgresapp.com/