[JRO] Small fixes after testing on a live data set.
Running the latest migration that splits posts into normal and private
posts bubbled up some small edge cases in the data for thredded.com.
There's an argument to be made for data normalization and enforcement of
data integrity at the main web app, but for the purposes of this commit,
we can guard against some of those integrities with the code itself.

1. There are some posts that have been orphaned by either the parent
   topic, or the parent user - and thus will break on
   `Post#update_parent_last_user_and_timestamp`. This commit adds a guard
   clause to ensure that we're updating an existing record with an existing
   user
2. Posts require *some* sort of content and will fail to save if content
   is empty. This update will add an ellipsis as the content if the content
   string is blank.