Refactor read state and improve performance
Read state is now handled the same way for private and regular topics.

Only loads the read state for the current user, resolving #243.

Adds a new (Private)Topic.unread(user) methods that returns a relation
with all unread topics. We can now show the count of unread topics in a
messageboard if we want to.

Fixes last page tracking.