[JRO] Round two times being compared down to ints
After trying things out with a live forum, it became apparent that there
was something going on with users' read tracking. After some digging I
found that the comparison `postable.updated_at <= read_at` was comparing
the timestamps down to the millisecond. So we were seeing some funny
race conditions based on when the updated_at and read_at columns were
being committed to the db. Example:

Thredded::Topic#updated_at            -> `1461352354.630764`

is effectively the same time as

Thredded::UserTopicReadStates#read_at -> `1461352354.615124`

So that comparison in the code *should* be intrepreted as truthy.
Instead, it's false.

Casting these two objects to integers fixes this just fine.