[JRO] Style "mark as read" button
In the process of moving the button around and styling it I've added an
additional "light" button variant to the stylesheet and the scss variables. Since I am not a designer these are subject to scrutiny and revision, of course.

Also, the resource originally created by @TheDudeWithTheThing was more a
*verb* than a noun so I've adjusted it to a more restful verb and made
the named path a more explicit `mark_all_private_topics_read` as this
action operates only on private topics.