Remove the &-- top-level class name trick (#418)
These serve no purpose and do not play well with %-placeholder
selectors. I initially added these thinking the `.thredded` wrapper
class name could be changed, but it can't and making it possible is
not worth the effort.

Additionally, this updates the following icons styles to %-selectors,
and adjusts their class names to conform to the naming scheme, i.e.:
`--topic-header--` for the topic posts view, and `--topics--` for
the topics list view.