Support Turbolinks 5
Turbolinks 5 caches the body by cloning it, which means none of the
bound event handlers are preserved (unlike Turbolinks classic).

This means all of the DOM transformations on load need to be either
idempotent, or the transformation should be reverted before caching.

This effectively means we need to call plugin destructors on
`turbolinks:before-cache`.

This PR adds support for Turbolinks 5, while preserving support for
no Turbolinks and Turbolinks Classic (with or without jquery-turbolinks).