Remove config.autoload_paths that are default
All of our config.autoload_paths are already Rails defaults for an
Engine, so we don't need to set them. We only need to set `lib` to
autoload.

Also, add a missing `require_dependency` for `thredded/errors`,
so that it's eager-loaded in production.

This also happens to improve compatibility with rails_admin's magic.
See #280 on why.