Do not use `require_dependency`
`require_dependency` is unnecessary, as per the findings in
https://github.com/thredded/thredded/issues/478#issuecomment-261743820

However, when not using `require_dependency`, the absolute constant name
must be used (e.g. `Thredded::Post` instead of `Post`) whenever
there is a chance that the constant has not been loaded yet.
This is because if the `Post` constant is already loaded, autoload will
not get triggered for `Thredded::Post` when referencing just `Post`.