Add Thredded.user_class_name, use in initializer
Previously we were calling `Thredded.user_class` which triggers auto-loading of the User class.

If the User class could not load (e.g. because there is a syntax error in the user.rb file),
the app would not load, which is contrary to the usual assumptions about running Rails in development.