Fix migrations for Rails 5.1
Rails 5.1 changes the default primary key type from integer to bigint.

This changes the upgrade migrations to create references that are of the
same type as they are referencing.

For the initial CreateThredded migration, the default PK type will be
used (so BIGINT on Rails 5.1, :integer on Rails < 5.1).

For v0.13 we will migrate all Thredded tables to use BIGINT primary keys.

This also removes upgrade migrations prior to v0.5 from the repository
to reduce maintenance burden.