Fix for module qualified user class
While trying to integrate Thredded with Spree I set Thredded.user_class to "Spree::User" which caused Thredded.user_class_name to be "spree/user" instead of "user" breaking most methods. This was the fix that seems to be working for me. I'm pretty new to Ruby/Rails so not sure if this is the best way to fix it.