Allow disabling all user links
This commit changes the user link partial to render a span element
instead of a link when the user_path lambda is evaluated to nil.

Therefor, it allows disabling all user links by setting the `user_path`
lambda to always return nil.

This can be useful if, for any reason, you only want to link to some
user profiles (just write the lambda so that it returns only valid
links, and nil otherwise), or if you want to completely disable the
profile link feature.