Feature: Spoiler tags
Spoiler tags can be added using <spoiler></spoiler> or
[spoiler][/spoiler] markup (configurable via
`Thredded::SpoilerTagFilter.spoiler_tags`).

Supported out of the box for any markup processor.
Spoilers are focusable and are activated on mousedown, spacebar, or enter.
They can also be nested.
Their accessibility role is "figure" and they have a dynamically updated
`aria-expanded` attribute.

In emails, a "select text to view" implementation is used instead, and
images are replaced with links.