Fix content caching on Rails 5.2
Thanks to @jflat06 for debugging this issue.
Fixes #712

Explanation from @jflat06:

The PostView object was acting as a proxy for Post caching. It included a function to forward
cache_key to the cache_key of Post.

However, in rails 5.2+, the defaults for how cache version have been updated, and now use a
different function, cache_key_with_version. This function doesn't exist in PostView, so the
new-style caching is getting confused and serving stale data. This commit adds forwarding for
cache_key_with_version from PostView->Post.