move failing spec into a specific directory
we need to have configuration specs so we don't need to generate
a separate dummy for each type of configuration spec.
However it isn't always possible to reload the dummy effectively between
specs so it's easier to just restart each one with a fresh context.
Each file can be separate change to the dummy